backend/requirements.txt
asgiref==3.8.1
gitpython==3.1.31 
pyan3==1.2.0
lizard==1.17.10
sqlparse==0.5.1
pyjwt==2.9.0
django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1